Implement patch:Get for single properties.
Let the plugin return values for single properties. This is useful for hosts
that query for properties individually. Hosts may want to query for
properties individually if there are really many properties and/or there is
interest for a limited number only (e.g. limited space on UI).

patch:Get and patch:Set queries are optionally checked for a matching
patch:subject, if a given host should have set it in the first place.